+++ contrib/platform/src/com/sun/jna/platform/win32/WinDef.java
@@ -54,7 +54,7 @@ public interface WinDef extends StdCallLibrary {
         }
 
         public DWORD(long value) {
-            super(4, value);
+            super(4, value, true);
         }
 		
         /**